fixed padding of species list in total trees
